+++ test/shpec/operation_shpec.sh
@@ -47,6 +47,7 @@ function __get_container_port ()
 # container - Docker container name.
 # counter - Timeout counter in seconds.
 # process_pattern - Regular expression pattern used to match running process.
+# ready_test - Command used to test if the service is ready.
 function __is_container_ready ()
 {
 	local container="${1:-}"
@@ -56,6 +57,7 @@ function __is_container_ready ()
 			'BEGIN { print 10 * seconds; }'
 	)
 	local process_pattern="${3:-}"
+	local ready_test="${4:-true}"
 
 	until (( counter == 0 )); do
 		sleep 0.1
@@ -63,7 +65,7 @@ function __is_container_ready ()
 		if docker exec ${container} \
 			bash -c "ps axo command \
 				| grep -qE \"${process_pattern}\" \
-				&& varnishadm vcl.show -v boot" \
+				&& eval \"${ready_test}\"" \
 			&> /dev/null
 		then
 			break
@@ -228,7 +230,8 @@ function test_basic_operations ()
 		if ! __is_container_ready \
 			varnish.pool-1.1.1 \
 			${STARTUP_TIME} \
-			"/usr/sbin/varnishd "
+			"/usr/sbin/varnishd " \
+			"varnishadm vcl.show -v boot"
 		then
 			exit 1
 		fi
@@ -648,7 +651,8 @@ function test_custom_configuration ()
 		if ! __is_container_ready \
 			varnish.pool-1.1.1 \
 			${STARTUP_TIME} \
-			"/usr/sbin/varnishd "
+			"/usr/sbin/varnishd " \
+			"varnishadm vcl.show -v boot"
 		then
 			exit 1
 		fi