+++ CHANGELOG.md
@@ -9,6 +9,7 @@ CentOS-6 6.9 x86_64 - Varnish Cache 4.1.
 ### 1.4.2 - Unreleased
 
 - Updates source image to [1.8.3 tag](https://github.com/jdeathe/centos-ssh/releases/tag/1.8.3).
+- Adds generic ready state test function.
 
 ### 1.4.1 - 2017-09-16